FAT-1889 Создан интерфейс IResponseBreedCallback<String, ImageView, DogKind> и его реализация в DogsKindAdapter'e
