+++ app/src/main/java/com/andersen/dogsapp/dogs/data/web/retrofitapi/IResponseBreedCallback.java
@@ -1,0 +1,6 @@
+package com.andersen.dogsapp.dogs.data.web.retrofitapi;
+
+public interface IResponseBreedCallback<String, ImageView, DogKind> {
+
+    void onResponseBreedCallbackListener(String breed, ImageView imageView, DogKind dogKind);
+}
