+++ app/src/main/res/layout/dog_kind_item.xml
@@ -24,8 +24,8 @@
 
             <ImageView
                 android:id="@+id/dog_kind_image_view"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
                 android:layout_centerInParent="true"/>
 
             <TextView
